feat: :sparkles: created endpoint /admin/data
